RF | Combining Pipeline classes (#2200)
RF | Combining Pipeline classes
