--- src/test/java/school/redrover/PipelineProjectTest.java
@@ -1,334 +1,0 @@
-package school.redrover;
-
-import com.github.javafaker.Faker;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.openqa.selenium.By;
-import org.openqa.selenium.WebElement;
-import org.openqa.selenium.support.Color;
-import org.openqa.selenium.support.ui.ExpectedConditions;
-import org.testng.Assert;
-import org.testng.annotations.Test;
-import school.redrover.runner.BaseTest;
-import school.redrover.runner.TestUtils;
-
-import java.util.*;
-
-public class PipelineProjectTest extends BaseTest {
-    final String EXPECTED_RESULT = "New pipeline project";
-
-    private void createWithoutDescription(String name) {
-        getDriver().findElement(By.xpath("//a[@href = 'newJob']")).click();
-
-        getDriver().findElement(By.id("name")).sendKeys(name);
-        getDriver().findElement(By.xpath("//*[@id='j-add-item-type-standalone-projects']//li[2]")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-
-        getDriver().findElement(By.name("Submit")).click();
-    }
-
-    public WebElement findElement(By by) {
-        return getDriver().findElement(by);
-    }
-
-    public void clickPageButton(String menuButton){
-        getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.xpath(String.format("//*[text()='%s']", menuButton)))).click();
-    }
-
-    public void clickTaskButton(String nameOfTask) {
-        List<WebElement> taskIconLinks = getDriver().findElements(By.className("task-icon-link"));
-        List<WebElement> taskLinksText = getDriver().findElements(By.xpath("//*[@id='tasks']/*//span[2]"));
-        List<String> textOfTaskLink = new ArrayList<>();
-        for (WebElement taskLinkText : taskLinksText) {
-            textOfTaskLink.add(taskLinkText.getText());
-        }
-
-        Map<WebElement, String> buttonAndTask = new HashMap<>();
-        int i = 0;
-        for (WebElement taskIcon : taskIconLinks) {
-            buttonAndTask.put(taskIconLinks.get(i), textOfTaskLink.get(i));
-            i++;
-        }
-
-        for (Map.Entry<WebElement, String> entry : buttonAndTask.entrySet()) {
-            if (entry.getValue().equals(nameOfTask)) {
-                entry.getKey().click();
-            } else {
-                System.out.println("Wrong item");
-            }
-        }
-    }
-
-    public void createPipelineProject(String nameOfProject, String typeOfProject) {
-        WebElement fieldName = findElement(By.id("name"));
-        getWait2().until(ExpectedConditions.visibilityOf(fieldName)).sendKeys(nameOfProject);
-
-        List<WebElement> listItemOptions = getDriver().findElements(By.id("j-add-item-type-standalone-projects"));
-        for (WebElement element : listItemOptions) {
-            if (element.getText().contains(typeOfProject)) {
-                element.click();
-            }
-        }
-        findElement(By.id("ok-button")).click();
-    }
-
-    public void clickButtonApply() {
-        findElement(By.name("Apply")).click();
-    }
-
-    public String statusOfProject() {
-        return getWait2().until(ExpectedConditions.visibilityOfElementLocated(
-                By.className("svg-icon"))).getAttribute("title");
-    }
-
-    public void disableProject(){
-        getWait5().until(ExpectedConditions.elementToBeClickable(By.id("toggle-switch-enable-disable-project"))).click();
-    }
-
-    @Test
-    public void testSetDescription() {
-        String descriptionText = "This is a test description";
-
-        createWithoutDescription("test-pipeline");
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        getDriver().findElement(By.name("description")).sendKeys(descriptionText);
-        getDriver().findElement(By.name("Submit")).click();
-
-        WebElement actualDescription = getDriver().findElement(By.xpath("//*[@id='description']/div"));
-
-        Assert.assertEquals(actualDescription.getText(), descriptionText);
-    }
-
-    @Test
-    public void testDiscardOldBuildsIsChecked() {
-        createWithoutDescription("test-pipeline");
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        getDriver().findElement(By.xpath("//label[contains(text(),'Discard old builds')]")).click();
-        getDriver().findElement(By.name("Submit")).click();
-
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        WebElement discardOldBuildsCheckbox = getDriver().findElement(By.id("cb2"));
-
-        Assert.assertTrue(discardOldBuildsCheckbox.isSelected());
-    }
-
-    @Test
-    public void testDiscardOldBuildsIsCheckedWithValidParams() {
-        final String days = "7";
-        final String builds = "5";
-
-        createWithoutDescription("test-pipeline");
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        getDriver().findElement(By.xpath("//label[contains(text(),'Discard old builds')]")).click();
-        getDriver().findElement(By.name("_.daysToKeepStr")).sendKeys(days);
-        getDriver().findElement(By.name("_.numToKeepStr")).sendKeys(builds);
-        getDriver().findElement(By.name("Submit")).click();
-
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        WebElement discardOldBuildsCheckbox = getDriver().findElement(By.id("cb2"));
-
-        Assert.assertTrue(discardOldBuildsCheckbox.isSelected());
-        Assert.assertEquals(getDriver().findElement(By.name("_.daysToKeepStr")).getAttribute("value"), days);
-        Assert.assertEquals(getDriver().findElement(By.name("_.numToKeepStr")).getAttribute("value"), builds);
-    }
-
-    @Test
-    public void testDiscardOldBuildsIsChecked0Days() {
-        final String days = "0";
-        final String errorMessage = "Not a positive integer";
-
-        createWithoutDescription("test-pipeline");
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        getDriver().findElement(By.xpath("//label[contains(text(),'Discard old builds')]")).click();
-        getDriver().findElement(By.name("_.daysToKeepStr")).sendKeys(days);
-
-        WebElement discardOldBuildsCheckbox = getDriver().findElement(By.id("cb2"));
-
-        WebElement daysToKeepLabel = getDriver()
-                .findElement(By.xpath("//*[@name='strategy']/div/div"));
-        daysToKeepLabel.click();
-
-        WebElement actualErrorMessage = getWait10().until(ExpectedConditions
-                .visibilityOfElementLocated(By.xpath("//*[@name='strategy']//div[@class='error']")));
-
-        Assert.assertTrue(discardOldBuildsCheckbox.isSelected());
-        Assert.assertEquals(actualErrorMessage.getText(), errorMessage);
-    }
-
-    @Test
-    public void testDiscardOldBuildsIsChecked0Builds() {
-        final String builds = "0";
-        final String errorMessage = "Not a positive integer";
-
-        createWithoutDescription("test-pipeline");
-        getDriver().findElement(By.xpath("//*[@href='/job/test-pipeline/configure']")).click();
-
-        getDriver().findElement(By.xpath("//label[contains(text(),'Discard old builds')]")).click();
-        getDriver().findElement(By.name("_.numToKeepStr")).sendKeys(builds);
-
-        WebElement discardOldBuildsCheckbox = getDriver().findElement(By.id("cb2"));
-
-        WebElement clickOutsideOfInputField = getDriver()
-                .findElement(By.xpath("//*[@name='strategy']/div/div"));
-        clickOutsideOfInputField.click();
-
-        WebElement actualErrorMessage = getWait5().until(ExpectedConditions
-                .visibilityOfElementLocated(By.xpath("//*[@name='strategy']//div[@class='error']")));
-
-        Assert.assertTrue(discardOldBuildsCheckbox.isSelected());
-        Assert.assertEquals(actualErrorMessage.getText(), errorMessage);
-    }
-
-    @Test
-    public void testDisableDuringCreation() {
-        final String PIPELINE_NAME = "My_pipeline";
-
-        getDriver().findElement(By.xpath("//a[@href='/view/all/newJob']")).click();
-
-        getWait2().until(ExpectedConditions.visibilityOf(getDriver().findElement(By.id("name")))).sendKeys(PIPELINE_NAME);
-        getDriver().findElement(By.cssSelector(".org_jenkinsci_plugins_workflow_job_WorkflowJob")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-
-        getWait5().until(ExpectedConditions.presenceOfElementLocated(By.id("toggle-switch-enable-disable-project")));
-
-        WebElement enableToggles = getDriver().findElement(By.id("toggle-switch-enable-disable-project"));
-        boolean isPipelineEnabled = Boolean.parseBoolean(getDriver().findElement(By.xpath("//input[@name='enable']")).getAttribute("value"));
-        if (isPipelineEnabled) {
-            enableToggles.click();
-        }
-
-        getDriver().findElement(By.name("Submit")).click();
-
-        getWait2().until(ExpectedConditions.textToBePresentInElement(getDriver().findElement(By.tagName("h1")), "Pipeline"));
-        String disabledWarning = getDriver().findElement(By.id("enable-project")).getText();
-
-        getDriver().findElement(By.xpath("//a[contains(@href,'configure')]")).click();
-
-        getWait5().until(ExpectedConditions.textToBe(By.tagName("h2"), "General"));
-        boolean isPipelineEnabledAfterDisable = Boolean.parseBoolean(getDriver().findElement(
-                By.xpath("//input[@name='enable']")).getAttribute("value"));
-
-        Assert.assertTrue(disabledWarning.contains("This project is currently disabled"));
-        Assert.assertFalse(isPipelineEnabledAfterDisable, "Pipeline is enabled");
-    }
-
-    @Test
-    public void addDescriptionPipelineProjectTest(){
-        String description = "This is a project for school test";
-        clickTaskButton("New Item");
-        createPipelineProject(EXPECTED_RESULT, "Pipeline");
-        findElement(By.name("description")).sendKeys(description);
-        clickButtonApply();
-        WebElement messageSaved = getWait2().until(ExpectedConditions.visibilityOfElementLocated(
-                By.id("notification-bar")));
-
-        Assert.assertTrue(messageSaved.isDisplayed());
-
-        clickPageButton("Dashboard");
-        clickPageButton(EXPECTED_RESULT);
-
-        WebElement fieldDescription = getWait2().until(ExpectedConditions.visibilityOfElementLocated(By.id("description")));
-
-        Assert.assertTrue(fieldDescription.getText().contains(description));
-    }
-
-    @Test(dependsOnMethods = "addDescriptionPipelineProjectTest")
-    public void disablePipelineProjectTest(){
-
-        clickPageButton("Dashboard");
-        String statusBeforeDisable = statusOfProject();
-
-        clickPageButton(EXPECTED_RESULT);
-        clickTaskButton("Configure");
-        disableProject();
-        clickButtonApply();
-
-        clickPageButton("Dashboard");
-        String statusAfterDisable = statusOfProject();
-
-        Assert.assertNotEquals(statusBeforeDisable, statusAfterDisable);
-    }
-
-    @Test(dependsOnMethods = "testDisableDuringCreation")
-    public void testSetDescription2() {
-        final String newDescription = "Pipeline description";
-
-        WebElement selectPipelineButton = getDriver()
-                .findElement(By.xpath("//a[@class='jenkins-table__link model-link inside']"));
-        selectPipelineButton.click();
-
-        WebElement addDescriptionButton = getWait2().until(ExpectedConditions
-                .elementToBeClickable(By.xpath("//a[@href='editDescription']")));
-        addDescriptionButton.click();
-
-        WebElement descriptionInput = getDriver().findElement(By.xpath("//textarea[@name='description']"));
-        descriptionInput.clear();
-        descriptionInput.sendKeys(newDescription);
-
-        WebElement saveButton = getDriver().findElement(By.xpath("//button[@formnovalidate='formNoValidate']"));
-        saveButton.click();
-
-        WebElement resultMessage = getWait2().until(ExpectedConditions
-                .elementToBeClickable(By.xpath("//div[@id='description']/div")));
-        String messageValue = resultMessage.getText();
-
-        Assert.assertEquals(messageValue, newDescription);
-    }
-
-    @Test
-    public void testCreatePipelineGoingFromManageJenkinsPage() {
-        final String RANDOM_NAME_PROJECT = RandomStringUtils.randomAlphanumeric(8);
-
-        getDriver().findElement(By.linkText("Manage Jenkins")).click();
-        getDriver().findElement(By.linkText("New Item")).click();
-        getWait5().until(ExpectedConditions.visibilityOfElementLocated(By.xpath("//input[@name='name']"))).sendKeys(RANDOM_NAME_PROJECT);
-        getDriver().findElement(By.xpath("//span[text()='Pipeline']")).click();
-        getDriver().findElement(By.id("ok-button")).click();
-        getDriver().findElement(By.name("Submit")).click();
-        getDriver().findElement(By.xpath("//ol//a[@href='/']")).click();
-
-        List<String> jobList = getDriver().findElements(By.cssSelector(".jenkins-table__link"))
-                .stream()
-                .map(WebElement::getText)
-                .toList();
-
-        Assert.assertTrue(jobList.contains(RANDOM_NAME_PROJECT));
-    }
-
-    @Test
-    public void buildNowFromPipelineView() {
-        String pipelineName = new Faker().name().title().replace(" ", "");
-        TestUtils.createPipeline(this, pipelineName, true);
-        getDriver().findElement(By.xpath("//*[@href='job/"+pipelineName+"/']")).click();
-
-        if (!getDriver().findElement(By.xpath("//div[@id='no-builds']")).isDisplayed()) {
-            getDriver().findElement(By.xpath("//a[@href='/toggleCollapse?paneId=buildHistory']")).click();
-        }
-
-        getDriver().findElement(By
-                .xpath("//a[@href='/job/"+pipelineName+"/build?delay=0sec']")).click();
-
-        int numberOfStartedBuilds = 1;
-        boolean lastBuildIsPresent = getWait5().until(ExpectedConditions.presenceOfElementLocated(By
-                        .xpath("//span[@class='build-status-icon__outer']//*[name()='svg']["+numberOfStartedBuilds+"]")))
-                .isDisplayed();
-        List<WebElement> list = new ArrayList<>();
-        if (lastBuildIsPresent) {
-            list = getDriver().findElements(By
-                    .xpath("//span[@class='build-status-icon__outer']//*[name()='svg']"));
-        }
-
-        try {
-            Assert.assertEquals(Color.fromString(list.get(0).getCssValue("color"))
-                    .asHex(), "#1ea64b");
-        } catch (NullPointerException | IndexOutOfBoundsException e) {
-            System.out.println("there are no builds in the 'Build History' list");
-            e.printStackTrace();
-        }
-    }
-}
